requirements.txt
streamlit>=1.30.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
plotly>=5.18.0
pandas>=2.1.0
tenacity>=8.2.0
weasyprint>=60.0
markdown>=3.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
